+++ src/collections/infra/repositories/transformers/collectionTransformers.ts
@@ -4,6 +4,7 @@ import {
   CollectionContactPayload,
   CollectionInputLevelPayload,
   CollectionPayload,
+  AllowedDatasetTypePayload
 } from './CollectionPayload'
 import { transformPayloadToOwnerNode } from '../../../../core/infra/repositories/transformers/dvObjectOwnerNodeTransformer'
 import { CollectionFacet } from '../../../domain/models/CollectionFacet'
@@ -13,7 +14,7 @@ import {
   CollectionItemSubset,
   CountPerObjectType
 } from '../../../domain/models/CollectionItemSubset'
-import { DatasetPreview } from '../../../../datasets'
+import { DatasetPreview, DatasetType } from '../../../../datasets'
 import { FilePreview } from '../../../../files'
 import { DatasetPreviewPayload } from '../../../../datasets/infra/repositories/transformers/DatasetPreviewPayload'
 import { FilePreviewPayload } from '../../../../files/infra/repositories/transformers/FilePreviewPayload'
@@ -85,6 +86,11 @@ const transformPayloadToCollection = (collectionPayload: CollectionPayload): Col
     }),
     ...(collectionPayload.dataverseContacts && {
       contacts: transformContactsPayloadToContacts(collectionPayload.dataverseContacts)
+    }),
+    ...(collectionPayload.allowedDatasetTypes && {
+      allowedDatasetTypes: transformAllowedDatasetTypesPayloadToAllowedDatasetTypes(
+        collectionPayload.allowedDatasetTypes
+      )
     })
   }
   return collectionModel
@@ -253,5 +259,16 @@ const transformContactsPayloadToContacts = (
   return contactsPayload.map((contactPayload) => ({
     email: contactPayload.contactEmail,
     displayOrder: contactPayload.displayOrder
+  }))
+}
+
+const transformAllowedDatasetTypesPayloadToAllowedDatasetTypes = (
+  allowedDatasetTypesPayload: AllowedDatasetTypePayload[]
+): DatasetType[] => {
+  return allowedDatasetTypesPayload.map((allowedDatasetType) => ({
+    id: allowedDatasetType.id,
+    name: allowedDatasetType.name,
+    displayName: allowedDatasetType.displayName,
+    description: allowedDatasetType.description
   }))
 }