+++ test/testHelpers/collections/collectionHelper.ts
@@ -48,6 +48,14 @@ export const createCollectionModel = (theme?: CollectionTheme): Collection => {
         displayOrder: 0
       }
     ],
+    allowedDatasetTypes: [
+      {
+        id: 1,
+        name: 'review',
+        displayName: 'Review',
+        description: 'A review of a dataset compiled by the expert community.'
+      }
+    ],
     isMetadataBlockRoot: true,
     isFacetRoot: true,
     childCount: 0,
@@ -77,6 +85,14 @@ export const createCollectionPayload = (theme?: CollectionThemePayload): Collect
       {
         contactEmail: 'dataverse@test.com',
         displayOrder: 0
+      }
+    ],
+    allowedDatasetTypes: [
+      {
+        id: 1,
+        name: 'review',
+        displayName: 'Review',
+        description: 'A review of a dataset compiled by the expert community.'
       }
     ],
     isMetadataBlockRoot: true,